PinotSegmentRestletResource done
