+++ pinot-controller/src/main/java/com/linkedin/pinot/controller/api/resources/SuccessResponse.java
@@ -15,9 +15,9 @@
  */
 package com.linkedin.pinot.controller.api.resources;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 
 public class SuccessResponse {
